Test for bugreport ID 2495252.
Report:
The test works by allocating two pages of virtual memory, and then using
mprotect to make the second page read-inaccessible. The first page is
still read-accessible, and we only try to read 10 characters from it, but
the strlen call searches past that for a zero-terminator and encounters the
read-inaccessible page, leading to a segfault.

Resolution: not confirmed, test pass fine.