GCC `-fvisibility=hidden' broken on SunOS through 4.3
It appears that the `-fvisibility=hidden' option to GCC causes GCC 4.0
through 4.3 on SunOS (Solaris) to generate defective .o files, which
the Sun linker chokes on. I think the bug is fixed in GCC 4.4.
The symptom is a large table with the heading "Text relocation remains
against symbol", followed by "ld: fatal: relocations remain against
allocatable but non-writable sections".

This is correspond to 3b085bb in STLport-5.2 branch.

Bug ID #2815529.