Fix tests: unordered containers are fine
<snip>
The elements of an unordered associative container
are organized into buckets. Keys with the same hash
code appear in the same bucket. The number of buckets
is automatically increased as elements are added
to an unordered associative container, so that
the average number of elements per bucket is kept below
a bound. Rehashing invalidates iterators, changes
ordering between elements, and changes which buckets
elements appear in, but does not invalidate pointers
or references to elements. For unordered_multiset
and unordered_multimap, rehashing preserves the relative
ordering of equivalent elements.
</snip>