Bug: erase in unordered container in loop
<snip>
The elements of an unordered associative container
are organized into buckets. Keys with the same hash
code appear in the same bucket. The number of buckets
is automatically increased as elements are added
to an unordered associative container, so that
the average number of elements per bucket is kept below
a bound.

<bold>
Rehashing invalidates iterators, changes
ordering between elements, and changes which buckets
elements appear in, but does not invalidate pointers
or references to elements.
</bold>

For unordered_multiset and unordered_multimap,
rehashing preserves the relative ordering
of equivalent elements.
</snip>

libxmt release 2.1.5.

libsockios release 2.6.10.

libstem release 5.4.4.

libjanus release 0.14.10.