Unordered containers: shouldn't rehash on erase
<snip>
The insert members shall not affect the validity of references
to container elements, but may invalidate all iterators
to the container. The erase members shall invalidate only
iterators and references to the erased elements.
</snip>

Rationale: if we would rehash on erase, we can't use
remove_if() for unordered containers.

Let's try to reduce amount of buckets on resize/rehash.