Port for ARM tool-chain.
The library is built  statically, because most of the armcc
users are in embedded world where the dynamic linking and/or
file-system don't exist, and that's how normally libraries
are used with the armcc compiler, including C++ libraries
shipped with the ARM toolchain. For others - statically
linked library with few limitations - still can be used. Dynamically
linked library is something will be considered in future.  The calls
to POSIX routines (mainly file-system) has been adjusted to make it
conforming to our ARM libraries.

Signed-off-by: Wojciech Daniel Meyer <wojciech.meyer@arm.com>