Reenable uncaught_exception() for ICL
Due to the way it was implemented (in the 'Fix _STLP_MSVC definition'
patch by dums on 2006-03-16), uncaught_exception() is disabled when
compiling with ICL.

Attached is a simple patch (for stlport 5.2.1) to reenable it.
This is only needed on MSWindows (intel compiler on linux uses
config/_icc.h instead of config/_msvc.h).