Fix is_constructible
Define

  template <class _Tp>
  _Tp&& __declval_transparent() _STLP_NOEXCEPT;

It useful when I need to return 'any' type, not
only rvalue (due to convertion-to-rvalue-reference rules).

This allow to check construction with lvalue, for example
in __is_assignable, the base for is_assignable.

This fix is_move_assignable checks for gcc 4.6.x
(looks, gcc 4.5 not worry about convertion-to-rvalue-reference rules
too much).